fix: use int64

+++ src/scenes/Result.hpp
@@ -6,7 +6,7 @@
 class Result : public App::Scene {
   constexpr const static char32_t* const scoresFile = U"scores.txt";
   Array<int32> scores;
-  uint32 rank;
+  int64 rank;
 public:
   Result(const InitData& init);
